Hack fix for crossover points in NLO kernels
For the "constant" terms arising from the plus prescription, as well as
the qG kernel, the code gives nans unless I push x away from the
DGLAP-ERBL boundary by a small amount. Ideally, it should be possible to
analytically solve for the kernels at these boundary points, but the NLO
kernels are complicated and this seems prohibitively difficult. Pushing
away from the boundary seems to work for now, though.